wellness-tracker-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Alembic==1.13.1
Flask-Migrate==4.0.5
